Prevent health and mana from being floats.

+++ game/world/managers/objects/units/creature/CreatureManager.py
@@ -136,8 +136,8 @@ class CreatureManager(UnitManager):
         self.spirit = creature_class_level_stats.spirit
         self.attack_power = creature_class_level_stats.attack_power
         self.ranged_attack_power = creature_class_level_stats.ranged_attack_power
-        self.max_health = creature_class_level_stats.health * creature_template.health_multiplier
-        self.max_power_1 = creature_class_level_stats.mana * creature_template.mana_multiplier
+        self.max_health = int(creature_class_level_stats.health * creature_template.health_multiplier)
+        self.max_power_1 = int(creature_class_level_stats.mana * creature_template.mana_multiplier)
         self.health = int((self.health_percent / 100) * self.max_health)
         self.power_1 = int((self.mana_percent / 100) * self.max_power_1)
